+++ src/config/plugin.ts
@@ -14,6 +14,7 @@ import {compact, exists, resolvePackage, flatMap, loadJSON, mapValues} from './u
 import {isProd, requireJson} from '../util'
 import ModuleLoader from '../module-loader'
 import {Command} from '../command'
+import {Performance} from '../performance'
 
 const _pjson = requireJson<PJSON>(__dirname, '..', '..', 'package.json')
 
@@ -124,6 +125,10 @@ export class Plugin implements IPlugin {
 
   children: Plugin[] = []
 
+  hasManifest = false
+
+  private _commandsDir!: string | undefined
+
   // eslint-disable-next-line new-cap
   protected _debug = Debug()
 
@@ -172,12 +177,16 @@ export class Plugin implements IPlugin {
   }
 
   public get commandsDir(): string | undefined {
-    return tsPath(this.root, this.pjson.oclif.commands)
+    if (this._commandsDir) return this._commandsDir
+
+    this._commandsDir = tsPath(this.root, this.pjson.oclif.commands, this.type)
+    return this._commandsDir
   }
 
   public get commandIDs(): string[] {
     if (!this.commandsDir) return []
 
+    const marker = Performance.mark(`plugin.commandIDs#${this.name}`, {plugin: this.name})
     this._debug(`loading IDs from ${this.commandsDir}`)
     const patterns = [
       '**/*.+(js|cjs|mjs|ts|tsx)',
@@ -192,6 +201,8 @@ export class Plugin implements IPlugin {
       return id === '' ? '.' : id
     })
     this._debug('found commands', ids)
+    marker?.addDetails({count: ids.length})
+    marker?.stop()
     return ids
   }
 
@@ -200,6 +211,7 @@ export class Plugin implements IPlugin {
   async findCommand(id: string, opts?: {must: boolean}): Promise<Command.Class | undefined>
 
   async findCommand(id: string, opts: {must?: boolean} = {}): Promise<Command.Class | undefined> {
+    const marker = Performance.mark(`plugin.findCommand#${this.name}.${id}`, {id, plugin: this.name})
     const fetch = async () => {
       if (!this.commandsDir) return
       const search = (cmd: any) => {
@@ -210,7 +222,7 @@ export class Plugin implements IPlugin {
 
       let m
       try {
-        const p = path.join(this.pjson.oclif.commands as string, ...id.split(':'))
+        const p = path.join(this.commandsDir ?? this.pjson.oclif.commands, ...id.split(':'))
         const {isESM, module, filePath} = await ModuleLoader.loadWithData(this, p)
         this._debug(isESM ? '(import)' : '(require)', filePath)
         m = module
@@ -228,7 +240,7 @@ export class Plugin implements IPlugin {
 
     const cmd = await fetch()
     if (!cmd && opts.must) error(`command ${id} not found`)
-
+    marker?.stop()
     return cmd
   }
 
@@ -241,6 +253,7 @@ export class Plugin implements IPlugin {
           process.emitWarning(`Mismatched version in ${this.name} plugin manifest. Expected: ${this.version} Received: ${manifest.version}\nThis usually means you have an oclif.manifest.json file that should be deleted in development. This file should be automatically generated when publishing.`)
         } else {
           this._debug('using manifest from', p)
+          this.hasManifest = true
           return manifest
         }
       } catch (error: any) {
@@ -252,12 +265,17 @@ export class Plugin implements IPlugin {
       }
     }
 
+    const marker = Performance.mark(`plugin.manifest#${this.name}`, {plugin: this.name})
     if (!ignoreManifest) {
       const manifest = await readManifest()
-      if (manifest) return manifest
+      if (manifest) {
+        marker?.addDetails({fromCache: true, commandCount: Object.keys(manifest.commands).length})
+        marker?.stop()
+        return manifest
+      }
     }
 
-    return {
+    const manifest = {
       version: this.version,
       commands: (await Promise.all(this.commandIDs.map(async id => {
         try {
@@ -274,6 +292,9 @@ export class Plugin implements IPlugin {
         return commands
       }, {} as {[k: string]: Command.Cached}),
     }
+    marker?.addDetails({fromCache: false, commandCount: Object.keys(manifest.commands).length})
+    marker?.stop()
+    return manifest
   }
 
   protected warn(err: string | Error | CLIError, scope?: string): void {